resort-platform/backend/requirements.txt
# -----------------------------
# Core Framework
# -----------------------------
fastapi==0.111.0
uvicorn[standard]==0.30.1

# -----------------------------
# Database & ORM
# -----------------------------
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# -----------------------------
# Data Validation & Settings
# -----------------------------
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# -----------------------------
# Authentication & Security
# -----------------------------
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# -----------------------------
# HTTP & Integrations
# -----------------------------
httpx==0.27.0

# -----------------------------
# Background Tasks & Utilities
# -----------------------------
tenacity==8.3.0
python-dateutil==2.9.0.post0

# -----------------------------
# Logging & Monitoring
# -----------------------------
loguru==0.7.2

# -----------------------------
# Development & Testing
# -----------------------------
pytest==8.2.2
pytest-asyncio==0.23.7
